Simplified decoding of results
This commit simplifies loading and converting results of a database operation.
The Decoder module now only generates Hashes, Arrays, or singular values. The
conversion into custom types is now done in the connection adapter.

This change is necessary to be able to preload associations, since we’ll need
some place to store these, and we cannot do this if we don’t know how to
write into the target type - or some target types might even be immutable.

This comes at the cost of no longer being able to “stream” query results
into a block - but this mode cannot support preloading associations. The
interface to “SQL.all” remains the same though, we “just” need to store
the records in an array temporarily.