Fixes association conversion
The previous release tried to convert **all** attributes when they
are arrays (has_many) or hashes (belongs_to/has_one) associations.

Not only has this been expensive, it also has been wrong (an attribute 
which is a hash in the database should not be converted), and it failed
for some queries.

This patch amends that.