simple/sql Properly deal w/incomplete models, w/custom type converters
If an object is passed in that implements a "from_complete_row" method, and the query
results are complete - i.e. all columns in the underlying table (via fq_table_name)
are in the resulting records - this mode uses the +into+ object's +from_complete_row+
method to convert the records. This mode is used by Simple::Store.

If the records are not complete, and contain more than a single value this mode behaves
like <tt>:immutable</tt>, i.e. they are read only.

If the records are not complete and contain only a single value this mode returns that
value for each record.