Reimplement locking
This uses the pg_*_xact_lock lock function - that once acquired, cannot be released,
but will release automatically once the transaction ends. This is safer than
the pg_advisory_lock group of functions.

For ease of use the key argument can also be a string - in which case we use
a hash function to derive a key value. Usage example:

     Simple::SQL.lock! "products", 12
     Simple::SQL.ask "UPDATE products SET ... WHERE id=12"

This commit also reimplements _locked(.., &block)_ based on _lock!_